Migrate `loungeMute` store to pinia
